update error toast (#115) and remove console logging
